+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Changed
+* Added
   * Added package json lint test.
 
 ## 4.5.0 - (July 21, 2021)