+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Added A11yLabel to ItemComment.
+  
 ## 4.6.0 - (March 29, 2023)
 
 * Changed