+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.7.0 - (May 15, 2023)
+
 * Deprecated
   * Deprecated `isDisabled` prop as the feature does not meet a11y standards.